Move logic for discovering BANDS HDU into utility function.
